+++ .gitignore
@@ -8,6 +8,7 @@
 /tmp/
 /spec/dummy/log
 /spec/dummy/tmp
+/spec/dummy/db/*.sqlite3*
 .env
 
 # rspec failure tracking
