+++ Gemfile
@@ -2,7 +2,3 @@ source "https://rubygems.org"
 
 # Specify your gem's dependencies in rails_api_logger.gemspec
 gemspec
-
-group :development, :test do
-  gem "minitest"
-end
