+++ app/models/rails_api_logger/inbound_request_log.rb
@@ -1,0 +1,5 @@
+module RailsApiLogger
+  class InboundRequestLog < RequestLog
+    self.table_name = "inbound_request_logs"
+  end
+end
