+++ app/models/rails_api_logger/outbound_request_log.rb
@@ -1,0 +1,5 @@
+module RailsApiLogger
+  class OutboundRequestLog < RequestLog
+    self.table_name = "outbound_request_logs"
+  end
+end
