+++ lib/rails_api_logger/engine.rb
@@ -1,9 +1,21 @@
-class RailsApiLogger
+module RailsApiLogger
   class Engine < ::Rails::Engine
     isolate_namespace RailsApiLogger
 
     config.generators do |g|
       g.test_framework :rspec
+    end
+
+    config.rails_api_logger = ActiveSupport::OrderedOptions.new
+
+    initializer "rails_api_logger.config" do
+      config.rails_api_logger.each do |name, value|
+        RailsApiLogger.public_send(:"#{name}=", value)
+      end
+    end
+
+    ActiveSupport.on_load(:action_controller) do
+      include InboundRequestsLogger
     end
   end
 end
