--- lib/rails_api_logger/inbound_request_log.rb
@@ -1,2 +1,0 @@
-class InboundRequestLog < RequestLog
-end
