--- lib/rails_api_logger/outbound_request_log.rb
@@ -1,2 +1,0 @@
-class OutboundRequestLog < RequestLog
-end
