+++ lib/rails_api_logger/version.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
-class RailsApiLogger
+module RailsApiLogger
   VERSION = "0.9.0"
 end
