--- spec/dummy/Gemfile
@@ -1,1 +1,0 @@
-# frozen_string_literal: true
