+++ spec/dummy/app/controllers/api/books_controller.rb
@@ -6,4 +6,10 @@ class Api::BooksController < ApplicationController
   def create
     head :no_content
   end
+
+  def update
+    @book = Book.find(params[:id])
+    @book.update!(title: nil)
+    head :no_content
+  end
 end
