--- spec/dummy/app/javascript/application.js
@@ -1,2 +1,0 @@
-// Configure your import map in config/importmap.rb. Read more: https://github.com/rails/importmap-rails
-import "controllers"
