+++ spec/dummy/app/models/book.rb
@@ -1,0 +1,2 @@
+class Book < ApplicationRecord
+end
