+++ spec/dummy/db/migrate/20241118205347_books.rb
@@ -1,0 +1,9 @@
+class Books < ActiveRecord::Migration[7.1]
+  def change
+    create_table :books do |t|
+      t.string :title, null: false
+      t.string :author, null: false
+      t.timestamps null: false
+    end
+  end
+end
