+++ spec/rails_api_logger_spec.rb
@@ -1,6 +1,6 @@
 RSpec.describe RailsApiLogger do
   it "defines some models" do
-    expect(InboundRequestLog.count).to eq(0)
-    expect(OutboundRequestLog.count).to eq(0)
+    expect(RailsApiLogger::InboundRequestLog.count).to eq(0)
+    expect(RailsApiLogger::OutboundRequestLog.count).to eq(0)
   end
 end
